end to end systems/2 - Chatbot with memory and Deployment/requirements.txt
openai==1.63.2
python-dotenv
uvicorn
fastapi